Drar ut VelgFil
